requirements.txt
beautifulsoup4==4.12.2
duckduckgo-search==3.8.5
faiss-cpu==1.7.4
gradio==3.42.0
huggingface-hub==0.16.4
langchain==0.0.276
langchain-experimental==0.0.12
openai==0.27.8
pydantic==2.3.0
requests==2.31.0
tiktoken==0.4.0
wikipedia==1.4.0
